requirements.txt
# Core Dependencies
Flask==2.3.3
Qiskit==1.0.0
numpy==1.24.3
python-dotenv==1.0.0

# Web Dependencies
Werkzeug==2.3.7
requests==2.31.0
websockets==11.0.3

# Security Dependencies
cryptography==41.0.3
pyOpenSSL==23.2.0

# File Handling
python-magic==0.4.27
Pillow==10.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.7.0
flake8==6.1.0
mypy==1.5.1

# Monitoring
prometheus-client==0.17.1
python-json-logger==2.0.7

# Documentation
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0 